Move saved pods from tappers provider to tappedPods provider
